refactor: apiWrapper 생성(로딩 및 에러메시지) 및 파일 분리
