+++ README.md.md
@@ -52,3 +52,4 @@
 - [ ] 전체 코드 리팩토링
   - [ ] 파일 및 컨트롤러 분리
   - [x] searchController 생성자 재생성 방지
+  - [x] api 호출시, 로딩스피너 및 에러메시지 공통처리
