--- src/api/getDetailMovieResult.ts
@@ -1,8 +1,0 @@
-import { MovieItemType } from "../types/movieResultType";
-import baseApi from "./baseApi";
-
-export const getDetailMovieResult = async (id: number): Promise<MovieItemType> => {
-  const movieResult = await baseApi(`/movie/${id}`);
-
-  return movieResult;
-};
